Consolidating UGens
