--- supriya/ugens/LFClipNoise.py
@@ -1,24 +1,0 @@
-import collections
-
-from supriya import CalculationRate
-from supriya.synthdefs import UGen
-
-
-class LFClipNoise(UGen):
-    """
-    A dynamic clipped noise generator.
-
-    ::
-
-        >>> supriya.ugens.LFClipNoise.ar()
-        LFClipNoise.ar()
-
-    """
-
-    ### CLASS VARIABLES ###
-
-    __documentation_section__ = "Noise UGens"
-
-    _ordered_input_names = collections.OrderedDict([("frequency", 500.0)])
-
-    _valid_calculation_rates = (CalculationRate.AUDIO, CalculationRate.CONTROL)
