--- supriya/ugens/Rotate2.py
@@ -1,40 +1,0 @@
-import collections
-
-from supriya import CalculationRate
-from supriya.synthdefs import MultiOutUGen
-
-
-class Rotate2(MultiOutUGen):
-    """
-    Equal-power sound-field rotator.
-
-    ::
-
-        >>> x = supriya.ugens.PinkNoise.ar() * 0.4
-        >>> y = supriya.ugens.LFTri.ar(frequency=880)
-        >>> y *= supriya.ugens.LFPulse.kr(frequency=3, width=0.1)
-        >>> position = supriya.ugens.LFSaw.kr(frequency=0.1)
-        >>> rotate_2 = supriya.ugens.Rotate2.ar(
-        ...     x=x,
-        ...     y=y,
-        ...     position=position,
-        ...     )
-        >>> rotate_2
-        UGenArray({2})
-
-    Returns an array of the rotator's left and right outputs.
-    """
-
-    ### CLASS VARIABLES ###
-
-    __documentation_section__ = "Spatialization UGens"
-
-    _default_channel_count = 2
-
-    _has_settable_channel_count = False
-
-    _ordered_input_names = collections.OrderedDict(
-        [("x", None), ("y", None), ("position", 0)]
-    )
-
-    _valid_calculation_rates = (CalculationRate.AUDIO, CalculationRate.CONTROL)
