--- supriya/ugens/XFade2.py
@@ -1,32 +1,0 @@
-import collections
-
-from supriya import CalculationRate
-from supriya.synthdefs import UGen
-
-
-class XFade2(UGen):
-    """
-    Two channel equal power crossfader.
-
-    ::
-
-        >>> xfade_3 = supriya.ugens.XFade2.ar(
-        ...     in_a=supriya.ugens.Saw.ar(),
-        ...     in_b=supriya.ugens.SinOsc.ar(),
-        ...     level=1,
-        ...     pan=supriya.ugens.LFTri.kr(frequency=0.1),
-        ...     )
-        >>> xfade_3
-        XFade2.ar()
-
-    """
-
-    ### CLASS VARIABLES ###
-
-    __documentation_section__ = None
-
-    _ordered_input_names = collections.OrderedDict(
-        [("in_a", None), ("in_b", 0), ("pan", 0), ("level", 1)]
-    )
-
-    _valid_calculation_rates = (CalculationRate.AUDIO, CalculationRate.CONTROL)
